remove dispacher module
