--- components/endpoint-configuration-management/org.wso2.carbon.identity.api.server.endpoint.configuration.dispatcher/src/main/webapp/META-INF/webapp-classloading.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!--
-  ~ Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ WSO2 Inc. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-
-
-<!--
-    This file defines class loading policy of the whole container. But this behaviour can be overridden by individual webapps by putting this file into the META-INF/ directory.
--->
-<Classloading xmlns="http://wso2.org/projects/as/classloading">
-
-    <!-- Parent-first or child-first. Default behaviour is child-first.-->
-    <ParentFirst>false</ParentFirst>
-
-    <!--
-	Default environments that contains provides to all the webapps. This can be overridden by individual webapps by specifing required environments
-	Tomcat environment is the default and every webapps gets it even if they didn't specify it.
-	e.g. If a webapps requires CXF, they will get both Tomcat and CXF.
-     -->
-    <Environments>CXF3,Carbon</Environments>
-</Classloading>
