--- components/endpoint-configuration-management/org.wso2.carbon.identity.api.server.endpoint.configuration.dispatcher/src/main/webapp/WEB-INF/beans.xml
@@ -1,56 +1,0 @@
-<!--
-  ~ Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-  ~
-  ~ WSO2 Inc. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-
-<beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:jaxrs="http://cxf.apache.org/jaxrs" xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:cxf="http://cxf.apache.org/core"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans  http://www.springframework.org/schema/beans/spring-beans-3.0.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd http://cxf.apache.org/jaxrs http://cxf.apache.org/schemas/jaxrs.xsd http://cxf.apache.org/core http://cxf.apache.org/schemas/core.xsd">
-    <import resource="classpath:META-INF/cxf/endpoint-management-server-v1-cxf.xml"/>
-    <context:property-placeholder/>
-    <context:annotation-config/>
-    <bean class="org.springframework.context.support.PropertySourcesPlaceholderConfigurer"/>
-    <bean class="org.springframework.beans.factory.config.PreferencesPlaceholderConfigurer"/>
-    <bean id="postprocess" class="org.springframework.beans.factory.annotation.AutowiredAnnotationBeanPostProcessor"/>
-    <bean id="validationProvider" class="org.apache.cxf.validation.BeanValidationProvider"/>
-    <bean id="validationInInterceptor"
-          class="org.apache.cxf.jaxrs.validation.JAXRSBeanValidationInInterceptor">
-        <property name="provider" ref="validationProvider"/>
-    </bean>
-    <jaxrs:server id="endpoint-configuration-management" address="/v1">
-        <jaxrs:serviceBeans>
-            <bean class="org.wso2.carbon.identity.api.server.endpoint.configuration.management.v1.EndpointConfigurationsApi"/>
-        </jaxrs:serviceBeans>
-        <jaxrs:providers>
-            <bean class="com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider">
-                <constructor-arg>
-                    <bean class="com.fasterxml.jackson.databind.ObjectMapper">
-                        <property name="serializationInclusion" value="NON_NULL" />
-                    </bean>
-                </constructor-arg>
-            </bean>
-            <bean class="com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider"/>
-            <bean class="org.wso2.carbon.identity.api.server.endpoint.configuration.dispatcher.JsonProcessingExceptionMapper"/>
-            <bean class="org.wso2.carbon.identity.api.server.endpoint.configuration.dispatcher.APIErrorExceptionMapper"/>
-            <bean class="org.wso2.carbon.identity.api.server.endpoint.configuration.dispatcher.InputValidationExceptionMapper"/>
-            <bean class="org.wso2.carbon.identity.api.server.endpoint.configuration.dispatcher.DefaultExceptionMapper"/>
-        </jaxrs:providers>
-        <jaxrs:inInterceptors>
-            <ref bean="validationInInterceptor"/>
-        </jaxrs:inInterceptors>
-    </jaxrs:server>
-</beans>
