+++ components/org.wso2.carbon.identity.api.server.endpoint.configuration.management/org.wso2.carbon.identity.api.server.endpoint.configuration.management.common/src/main/java/org/wso2/carbon/identity/api/server/endpoint/configuration/management/common/ContextLoader.java
