+++ components/org.wso2.carbon.identity.api.server.endpoint.configuration.management/org.wso2.carbon.identity.api.server.endpoint.configuration.management.v1/src/main/resources/META-INF/cxf/endpoint-management-server-v1-cxf.xml
